Fix test output directory
Instead of writing it to the current directory, write test output to target/tests/ directory.

Fixes #335.
