+++ cekit/tools.py
@@ -83,3 +83,16 @@ def get_brew_url(md5):
                      ex.output)
         raise ex
     return url
+
+class Chdir(object):
+    """ Context manager for changing the current working directory """
+
+    def __init__(self, newPath):
+        self.newPath = os.path.expanduser(newPath)
+
+    def __enter__(self):
+        self.savedPath = os.getcwd()
+        os.chdir(self.newPath)
+
+    def __exit__(self, etype, value, traceback):
+        os.chdir(self.savedPath)
