+++ tests/test_expose_services.py
@@ -9,7 +9,7 @@ import socket
 import sys
 import yaml
 
-from cekit.builders.osbs import Chdir
+from cekit.tools import Chdir
 from cekit.cli import Cekit
 
 image_descriptor = {
